Refactored report and utc time to run report queries
