+++ pstatus-notifications-workflow-ktor/src/main/kotlin/query/UtcTimeToRunReportQuery.kt
@@ -1,6 +1,7 @@
 package gov.cdc.ocio.processingnotifications.query
 
 import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
+import gov.cdc.ocio.types.InstantRange
 import java.time.LocalDate
 
 
@@ -35,5 +36,30 @@ abstract class UtcTimeToRunReportQuery(
         override fun build(): ReportQuery {
             throw NotImplementedError("This function must be implemented by derived classes")
         }
+    }
+
+    /**
+     * Use this to append the where class that filters by the data stream ids, data stream routes, jurisdictions,
+     * and date provided.
+     *
+     * @param utcDateToRun LocalDate
+     * @return String
+     */
+    fun whereClause(
+        utcDateToRun: LocalDate
+    ): String {
+        val instantRange = InstantRange.fromLocalDate(utcDateToRun)
+        val startTime = timeFunc(instantRange.start.epochSecond)
+        val endTime = timeFunc(instantRange.endInclusive.epochSecond)
+
+        val querySB = StringBuilder()
+        querySB.append(super.whereClause())
+
+        querySB.append("""
+                AND ${cPrefix}dexIngestDateTime >= $startTime
+                AND ${cPrefix}dexIngestDateTime < $endTime
+                """)
+
+        return querySB.toString().trimIndent()
     }
 }
